Moved dummy data into seperate file
