+++ src/App.js
@@ -1,7 +1,12 @@
+import Home from './components/Home'
+
 
 function App() {
   return (
+    <>
    <h1>Welcome to Trip Traveler's Trusty Transit</h1>
+   <Home />
+   </>
   );
 }
