+++ src/dummy_data.js
@@ -1,0 +1,40 @@
+const dummyData = [
+    {
+      id: 1,
+      name: "Coffee Shop",
+      city: "New York",
+      image: "https://images.unsplash.com/photo-1453614512568-c4024d13c247?ixid=MnwxMjA3fDB8MHxzZWFyY2h8MXx8Y29mZmVlJTIwc2hvcHxlbnwwfHwwfHw%3D&ixlib=rb-1.2.1&w=1000&q=80",
+      averageRating: 4,
+      pricePoint: "$2-$12",
+      reviews:9
+    },
+    {
+      id: 2,
+      name: "Coffee Shop",
+      city: "New York",
+      image: "https://images.unsplash.com/photo-1453614512568-c4024d13c247?ixid=MnwxMjA3fDB8MHxzZWFyY2h8MXx8Y29mZmVlJTIwc2hvcHxlbnwwfHwwfHw%3D&ixlib=rb-1.2.1&w=1000&q=80",
+      averageRating: 3,
+      pricePoint: "$2-$12",
+      reviews:7
+    },
+    {
+      id:3,
+      name: "Coffee Shop",
+      city: "New York",
+      image: "https://images.unsplash.com/photo-1453614512568-c4024d13c247?ixid=MnwxMjA3fDB8MHxzZWFyY2h8MXx8Y29mZmVlJTIwc2hvcHxlbnwwfHwwfHw%3D&ixlib=rb-1.2.1&w=1000&q=80",
+      averageRating: 1,
+      pricePoint: "$2-$12",
+      reviews:4
+    },
+    {
+      id:4,
+      name: "Coffee Shop",
+      city: "New York",
+      image: "https://images.unsplash.com/photo-1453614512568-c4024d13c247?ixid=MnwxMjA3fDB8MHxzZWFyY2h8MXx8Y29mZmVlJTIwc2hvcHxlbnwwfHwwfHw%3D&ixlib=rb-1.2.1&w=1000&q=80",
+      averageRating: 1,
+      pricePoint: "$2-$12",
+      reviews:4
+    },
+  ]
+
+export default dummyData
